refactor(cli): migrate nwp tooling onto discovery
Move NWP's .idx prober, the per-backend deep availability sampler, the offline
validator, and the direct-https live validator (plus the Herbie template / idx
helpers) to earthlens.nwp.cli. Relocate the CLI tests. Shrink the guard allow-list
(nwp) — only ecmwf remains.
