+++ libs/core/tests/cli/test_curate.py
@@ -85,49 +85,6 @@ class TestEcmwfProbe:
         assert sorted(result.assets) == ["2m_temperature", "tp"], "vars unioned"
 
 
-class TestNwpProbe:
-    """Tests for the NWP `.idx` band prober (Herbie template, no eccodes)."""
-
-    def test_reports_band_presence(self, monkeypatch):
-        """nwp probe flags which catalog band tokens appear in the live .idx."""
-        from earthlens.cli.adapter import load_catalog
-
-        catalog = load_catalog(_info("nwp"))
-        model_key = next(
-            key
-            for key, model in catalog.datasets.items()
-            if getattr(model, "model_family", None)
-            not in curate_mod._NWP_NO_IDX_FAMILIES | curate_mod._NWP_NEEDS_EXTRA_ATTRS
-            and (getattr(model, "bands", None) or {})
-        )
-        token = next(iter(catalog.datasets[model_key].bands.values()))
-        monkeypatch.setattr(
-            curate_mod, "_nwp_idx_body", lambda model: f"1:0:d=x:{token}:surface:\n"
-        )
-        result = probe_dataset(_info("nwp"), model_key)
-        assert result.status == "ok", "nwp probe ran"
-        assert any(v["present"] for v in result.assets.values()), "a band present"
-
-    def test_no_idx_family_is_error(self):
-        """An ECCC model (no .idx) reports 'error' with the reason."""
-        from earthlens.cli.adapter import load_catalog
-
-        catalog = load_catalog(_info("nwp"))
-        eccc = next(
-            (
-                key
-                for key, model in catalog.datasets.items()
-                if getattr(model, "model_family", None)
-                in curate_mod._NWP_NO_IDX_FAMILIES
-            ),
-            None,
-        )
-        if eccc is None:
-            pytest.skip("no ECCC model in the catalog")
-        result = probe_dataset(_info("nwp"), eccc)
-        assert result.status == "error" and "no .idx" in result.detail
-
-
 class TestDeepProbers:
     """Tests for the credentialed `--deep` samplers (creds/network mocked)."""
 
@@ -149,62 +106,6 @@ class TestDeepProbers:
         monkeypatch.setattr(stac_cli, "get_json", lambda url: _SAMPLE_ITEM)
         result = probe_dataset(_info("stac"), "sentinel-2-l2a", deep=True)
         assert result.status == "ok", "stac --deep fell back to the light prober"
-
-    def test_nwp_availability_direct_https_builds_url(self, monkeypatch):
-        """_nwp_availability HEADs the first band's URL for a direct-https model."""
-        import datetime as dt
-
-        from earthlens.nwp.catalog import NWPModel
-
-        calls = {}
-
-        class _Resp:
-            status_code = 200
-
-        def fake_head(url, timeout=None, allow_redirects=None):
-            calls["url"] = url
-            return _Resp()
-
-        monkeypatch.setattr(curate_mod.requests, "head", fake_head)
-        model = NWPModel(
-            provider="dwd-opendata",
-            backend="direct-https",
-            cycles_utc=[0],
-            url_template="https://x/{var_lc}/f{step:03d}_{var}.bz2",
-            bands={"temperature_2m": "T_2M"},
-        )
-        result = curate_mod._nwp_availability(model, dt.datetime(2024, 6, 1, 0), 0)
-        assert "HTTP 200" in result and calls["url"] == "https://x/t_2m/f000_T_2M.bz2"
-
-    def test_nwp_availability_herbie_unavailable(self, monkeypatch):
-        """_nwp_availability reports herbie missing rather than raising."""
-        import datetime as dt
-        import sys
-
-        from earthlens.nwp.catalog import NWPModel
-
-        monkeypatch.setitem(sys.modules, "herbie", None)
-        model = NWPModel(provider="noaa-nodd", model_family="gfs", backend="herbie")
-        result = curate_mod._nwp_availability(model, dt.datetime(2024, 6, 1, 0), 0)
-        assert "herbie unavailable" in result
-
-    def test_nwp_deep_reports_live_availability(self, monkeypatch):
-        """nwp --deep reports the model's live availability for a recent cycle."""
-        from earthlens.cli.adapter import load_catalog
-
-        monkeypatch.setattr(
-            curate_mod, "_nwp_availability", lambda model, cycle, step: "HTTP 200 (ok)"
-        )
-        catalog = load_catalog(_info("nwp"))
-        model_key = next(
-            key
-            for key, model in catalog.datasets.items()
-            if getattr(model, "backend", None) == "direct-https"
-        )
-        result = probe_dataset(_info("nwp"), model_key, deep=True)
-        assert result.status == "ok", "nwp deep probe ran"
-        entry = next(iter(result.assets.values()))
-        assert "HTTP 200" in entry["status"], "availability status reported"
 
 
 class TestProbeResult:
@@ -309,91 +210,6 @@ class TestGhslProberBranches:
         """An unknown GHSL product reports 'error'."""
         result = probe_dataset(_info("ghsl"), "not-a-ghsl-product")
         assert result.status == "error", "unknown product -> error"
-
-
-class TestNwpHelpers:
-    """Tests for the NWP availability dispatch + cycle helpers."""
-
-    def test_recent_cycle_is_in_the_past(self):
-        """_nwp_recent_cycle returns a datetime at or before ~now."""
-        import datetime as dt
-
-        from earthlens.nwp.catalog import NWPModel
-
-        cycle = curate_mod._nwp_recent_cycle(
-            NWPModel(provider="p", backend="direct-https", cycles_utc=[0, 12])
-        )
-        assert cycle <= dt.datetime.now(dt.UTC).replace(tzinfo=None), "cycle in past"
-
-    def test_availability_unknown_backend(self):
-        """An unrecognised backend reports that no probe exists."""
-        import datetime as dt
-        from types import SimpleNamespace
-
-        model = SimpleNamespace(backend="mystery", bands={}, request_options={})
-        out = curate_mod._nwp_availability(model, dt.datetime(2024, 1, 1), 0)
-        assert "no live availability probe" in out, "unknown backend reported"
-
-    def test_availability_direct_boto3_head_object(self, monkeypatch):
-        """A direct-boto3 model HEADs the object and reports its size."""
-        import datetime as dt
-        import sys
-        import types
-
-        from earthlens.nwp.catalog import NWPModel
-
-        boto3 = types.ModuleType("boto3")
-        botocore = types.ModuleType("botocore")
-        botocore_client = types.ModuleType("botocore.client")
-        botocore.UNSIGNED = object()
-        botocore_client.Config = lambda **kw: None
-
-        class FakeS3:
-            def head_object(self, Bucket, Key):
-                return {"ContentLength": 1234}
-
-        boto3.client = lambda *a, **kw: FakeS3()
-        monkeypatch.setitem(sys.modules, "boto3", boto3)
-        monkeypatch.setitem(sys.modules, "botocore", botocore)
-        monkeypatch.setitem(sys.modules, "botocore.client", botocore_client)
-        model = NWPModel(
-            provider="p",
-            backend="direct-boto3",
-            bands={"t": "T"},
-            request_options={"bucket": "b", "key_template": "{var}.grib"},
-        )
-        out = curate_mod._nwp_availability(model, dt.datetime(2024, 1, 1), 0)
-        assert "1234 bytes" in out, "head_object size reported"
-
-
-class TestNwpAvailabilityBranches:
-    """Branch coverage for the NWP availability probe."""
-
-    def test_availability_meteofrance_needs_key(self, monkeypatch):
-        """A meteofrance model with no API key reports the missing-credential."""
-        import datetime as dt
-
-        from earthlens.nwp.catalog import NWPModel
-
-        monkeypatch.delenv("METEO_FRANCE_API_KEY", raising=False)
-        monkeypatch.delenv("MF_API_KEY", raising=False)
-        model = NWPModel(
-            provider="mf",
-            backend="meteofrance-api",
-            request_options={"api_base": "https://x", "coverage_service": "svc"},
-        )
-        out = curate_mod._nwp_availability(model, dt.datetime(2024, 1, 1), 0)
-        assert "METEO_FRANCE_API_KEY" in out, "missing key reported"
-
-    def test_availability_direct_boto3_missing_options(self):
-        """A direct-boto3 model lacking bucket/key/bands reports the gap."""
-        import datetime as dt
-
-        from earthlens.nwp.catalog import NWPModel
-
-        model = NWPModel(provider="p", backend="direct-boto3")
-        out = curate_mod._nwp_availability(model, dt.datetime(2024, 1, 1), 0)
-        assert "bucket" in out, "missing options reported"
 
 
 class TestDeepSamplers:
@@ -509,144 +325,6 @@ class TestDeepSamplers:
         monkeypatch.setitem(sys.modules, "netCDF4", types.ModuleType("netCDF4"))
         monkeypatch.setattr(curate_mod, "_ecmwf_constraints", lambda d: [])
         assert curate_mod._ecmwf_deep_sample("x") == {}
-
-
-class TestNwpAvailabilityBackends:
-    """Cover the remaining _nwp_availability backend branches (SDKs faked)."""
-
-    def test_ecmwf_opendata_latest(self, monkeypatch):
-        """The ecmwf-opendata backend reports the latest cycle from the client."""
-        import datetime as dt
-        import sys
-        import types
-
-        from earthlens.nwp.catalog import NWPModel
-
-        opendata = types.ModuleType("ecmwf.opendata")
-        opendata.Client = lambda source=None, model=None: types.SimpleNamespace(
-            latest=lambda **kw: dt.datetime(2024, 6, 1, 0)
-        )
-        monkeypatch.setitem(sys.modules, "ecmwf", types.ModuleType("ecmwf"))
-        monkeypatch.setitem(sys.modules, "ecmwf.opendata", opendata)
-        model = NWPModel(
-            provider="p", backend="ecmwf-opendata", request_options={"type": "fc"}
-        )
-        out = curate_mod._nwp_availability(model, dt.datetime(2024, 1, 1), 0)
-        assert "latest cycle" in out, "latest cycle reported"
-
-    def test_herbie_resolves_grib(self, monkeypatch):
-        """The herbie backend reports the resolved GRIB path."""
-        import datetime as dt
-        import sys
-        import types
-
-        from earthlens.nwp.catalog import NWPModel
-
-        herbie = types.ModuleType("herbie")
-        herbie.Herbie = lambda cycle, **kw: types.SimpleNamespace(grib="s3://x.grib")
-        monkeypatch.setitem(sys.modules, "herbie", herbie)
-        model = NWPModel(provider="p", backend="herbie", model_family="gfs")
-        out = curate_mod._nwp_availability(model, dt.datetime(2024, 1, 1), 0)
-        assert "resolved" in out, "GRIB path reported"
-
-    def test_direct_https_unreachable(self, monkeypatch):
-        """A direct-https HEAD failure is reported as unreachable, not raised."""
-        import datetime as dt
-
-        from earthlens.nwp.catalog import NWPModel
-
-        def boom(url, timeout=None, allow_redirects=None):
-            raise RuntimeError("dns")
-
-        monkeypatch.setattr(curate_mod.requests, "head", boom)
-        model = NWPModel(
-            provider="p",
-            backend="direct-https",
-            url_template="https://x/{var}",
-            bands={"t": "T"},
-        )
-        out = curate_mod._nwp_availability(model, dt.datetime(2024, 1, 1), 0)
-        assert "unreachable" in out, "HEAD failure reported"
-
-
-class TestNwpIdx:
-    """Cover the Herbie `.idx` URL + body helpers (runpy / requests mocked)."""
-
-    def test_idx_url_from_template(self, monkeypatch):
-        """_nwp_idx_url evaluates the template against a stub to recover the URL."""
-        import datetime as dt
-        import pathlib
-        import runpy
-
-        from earthlens.nwp.catalog import NWPModel
-
-        class _Tmpl:
-            @staticmethod
-            def template(stub):
-                stub.SOURCES = {"aws": "https://aws/file"}
-
-        monkeypatch.setattr(runpy, "run_path", lambda p: {"gfs": _Tmpl})
-        model = NWPModel(
-            provider="p", backend="direct-https", model_family="gfs", product=""
-        )
-        url = curate_mod._nwp_idx_url(
-            pathlib.Path("/x"), model, dt.datetime(2024, 1, 1), 0
-        )
-        assert url == "https://aws/file.idx", "aws source + .idx suffix"
-
-    def test_idx_body_returns_reachable_text(self, monkeypatch):
-        """_nwp_idx_body returns the first reachable cycle's .idx text."""
-        import pathlib
-        import types
-
-        from earthlens.nwp.catalog import NWPModel
-
-        monkeypatch.setattr(
-            curate_mod, "_herbie_models_dir", lambda: pathlib.Path("/x")
-        )
-        monkeypatch.setattr(curate_mod, "_nwp_idx_url", lambda md, m, c, s: "https://x")
-        monkeypatch.setattr(
-            curate_mod.requests,
-            "get",
-            lambda url, timeout=None: types.SimpleNamespace(
-                status_code=200, text="1:0:VAR:\n"
-            ),
-        )
-        model = NWPModel(
-            provider="p", backend="direct-https", horizon_h=6, bands={"t": "VAR"}
-        )
-        assert curate_mod._nwp_idx_body(model) == "1:0:VAR:\n", "idx text returned"
-
-    def test_idx_body_unreachable_raises(self, monkeypatch):
-        """When no cycle is reachable, _nwp_idx_body raises ValueError."""
-        import pathlib
-
-        from earthlens.nwp.catalog import NWPModel
-
-        monkeypatch.setattr(
-            curate_mod, "_herbie_models_dir", lambda: pathlib.Path("/x")
-        )
-        monkeypatch.setattr(curate_mod, "_nwp_idx_url", lambda md, m, c, s: "https://x")
-
-        def boom(url, timeout=None):
-            raise RuntimeError("offline")
-
-        monkeypatch.setattr(curate_mod.requests, "get", boom)
-        model = NWPModel(provider="p", backend="direct-https", bands={"t": "VAR"})
-        with pytest.raises(ValueError, match="no recent"):
-            curate_mod._nwp_idx_body(model)
-
-    def test_idx_url_rejects_unsafe_model_family(self):
-        """A model_family that is not a bare identifier is refused before runpy."""
-        import pathlib
-
-        from earthlens.nwp.catalog import NWPModel
-
-        model = NWPModel(
-            provider="p", backend="direct-https", model_family="../evil", product=""
-        )
-        with pytest.raises(ValueError, match="unsafe model_family"):
-            curate_mod._nwp_idx_url(pathlib.Path("/x"), model, "2024-01-01", 0)
 
 
 class TestBiodiversityProbers:
